[Android][Project][Follow WE]
1. Add navigation bar
2. Add distance count and date count -> Reasonable walking speed
3. Add class UserRoute which can be saved & restored

Issue
Serializable used for bundle as saving and restore but Google class LatLng do not implements Serializable
We need to declaed as transient and do it on our module.
Moreover, it also cause item can not be JSONed.
Need to firgure out the better way

Title:
Type:
Root Cause:
Solution:
Feature:
ToDO: